+++ templates/add_edit_account.html
@@ -29,7 +29,7 @@
                             <button type="submit" class="btn btn-primary">
                                 <i class="fas fa-save me-1"></i>Save Account
                             </button>
-                            <a href="{{ url_for('accounts') }}" class="btn btn-outline-secondary">Cancel</a>
+                            <a href="{{ url_for('acct.accounts') }}" class="btn btn-outline-secondary">Cancel</a>
                         </div>
                     </form>
                 </div>
