+++ templates/add_edit_budget.html
@@ -49,7 +49,7 @@
                             <button type="submit" class="btn btn-primary">
                                 <i class="fas fa-save me-1"></i>Save Budget
                             </button>
-                            <a href="{{ url_for('budgets') }}" class="btn btn-outline-secondary">Cancel</a>
+                            <a href="{{ url_for('budget.budgets') }}" class="btn btn-outline-secondary">Cancel</a>
                         </div>
                     </form>
                 </div>
