+++ templates/add_edit_cashflow.html
@@ -46,7 +46,7 @@
                             <button type="submit" class="btn btn-primary">
                                 <i class="fas fa-save me-1"></i>Save cashflow
                             </button>
-                            <a href="{{ url_for('cashflows') }}" class="btn btn-outline-secondary">Cancel</a>
+                            <a href="{{ url_for('cashflow.cashflows') }}" class="btn btn-outline-secondary">Cancel</a>
                         </div>
                     </form>
                 </div>
