+++ templates/add_edit_category.html
@@ -31,7 +31,7 @@
                             <button type="submit" class="btn btn-primary">
                                 <i class="fas fa-save me-1"></i>Save Category
                             </button>
-                            <a href="{{ url_for('categories') }}" class="btn btn-outline-secondary">Cancel</a>
+                            <a href="{{ url_for('category.categories') }}" class="btn btn-outline-secondary">Cancel</a>
                         </div>
                     </form>
                 </div>
