+++ templates/add_edit_transaction.html
@@ -52,7 +52,7 @@
                             <button type="submit" class="btn btn-primary">
                                 <i class="fas fa-save me-1"></i>Save Transaction
                             </button>
-                            <a href="{{ url_for('transactions') }}" class="btn btn-outline-secondary">Cancel</a>
+                            <a href="{{ url_for('transact.transactions') }}" class="btn btn-outline-secondary">Cancel</a>
                         </div>
                     </form>
                 </div>
