+++ templates/budgets.html
@@ -6,7 +6,7 @@
 <div class="main-content">
     <div class="d-flex justify-content-between align-items-center mb-4">
         <h1><i class="fas fa-chart-pie me-2"></i>Budgets</h1>
-        <a href="{{ url_for('add_budget') }}" class="btn btn-primary">
+        <a href="{{ url_for('budget.add_budget') }}" class="btn btn-primary">
             <i class="fas fa-plus me-1"></i>Add Budget
         </a>
     </div>
@@ -96,7 +96,7 @@
                         <i class="fas fa-chart-pie fa-3x text-muted mb-3"></i>
                         <h5>No budgets found</h5>
                         <p class="text-muted">Create budgets for {{ datetime(year, month, 1).strftime('%B %Y') }}.</p>
-                        <a href="{{ url_for('add_budget') }}" class="btn btn-primary">
+                        <a href="{{ url_for('budget.add_budget') }}" class="btn btn-primary">
                             <i class="fas fa-plus me-1"></i>Add Budget
                         </a>
                     </div>
