+++ templates/cashflows.html
@@ -7,7 +7,7 @@
 <div class="main-content">
     <div class="d-flex justify-content-between align-items-center mb-4">
         <h1><i class="fas fa-adjust"></i> cashflows</h1>
-        <a href="{{ url_for('add_cashflow') }}" class="btn btn-primary">
+        <a href="{{ url_for('cashflow.add_cashflow') }}" class="btn btn-primary">
             <i class="fas fa-plus me-1"></i>Add cashflow
         </a>
     </div>
@@ -49,7 +49,7 @@
                             {% for cashflow in cashflows %}
                                 <tr>
                                     <td>
-                                        {% set edit_url = url_for('edit_cashflow') %}
+                                        {% set edit_url = url_for('cashflow.edit_cashflow') %}
                                         {% set edit_value = None %}
                                         {% include 'edit_button.html' %}
                                     </td>
@@ -83,7 +83,7 @@
                     <ul class="pagination justify-content-center">
                         {% if has_prev %}
                             <li class="page-item">
-                                <a class="page-link" href="{{ url_for('cashflows', page=page-1) }}">Previous</a>
+                                <a class="page-link" href="{{ url_for('cashflow.cashflows', page=page-1) }}">Previous</a>
                             </li>
                         {% else %}
                             <li class="page-item disabled">
@@ -97,7 +97,7 @@
                         
                         {% if has_next %}
                             <li class="page-item">
-                                <a class="page-link" href="{{ url_for('cashflows', page=page+1) }}">Next</a>
+                                <a class="page-link" href="{{ url_for('cashflow.cashflows', page=page+1) }}">Next</a>
                             </li>
                         {% else %}
                             <li class="page-item disabled">
@@ -111,7 +111,7 @@
                     <i class="fas fa-adjust"></i>
                     <h5>No cashflows found</h5>
                     <p class="text-muted">Start by adding your first cashflow.</p>
-                    <a href="{{ url_for('add_cashflow') }}" class="btn btn-primary">
+                    <a href="{{ url_for('cashflow.add_cashflow') }}" class="btn btn-primary">
                         <i class="fas fa-plus me-1"></i>Add cashflow
                     </a>
                 </div>
