+++ templates/categories.html
@@ -7,7 +7,7 @@
 <div class="main-content">
     <div class="d-flex justify-content-between align-items-center mb-4">
         <h1><i class="fas fa-tags me-2"></i>Categories</h1>
-        <a href="{{ url_for('add_category') }}" class="btn btn-primary">
+        <a href="{{ url_for('category.add_category') }}" class="btn btn-primary">
             <i class="fas fa-plus me-1"></i>Add Category
         </a>
     </div>
@@ -64,7 +64,7 @@
                         <i class="fas fa-tags fa-3x text-muted mb-3"></i>
                         <h5>No categories found</h5>
                         <p class="text-muted">Get started by adding income and expense categories.</p>
-                        <a href="{{ url_for('add_category') }}" class="btn btn-primary">
+                        <a href="{{ url_for('category.add_category') }}" class="btn btn-primary">
                             <i class="fas fa-plus me-1"></i>Add Category
                         </a>
                     </div>
