+++ templates/dashboard.html
@@ -7,7 +7,7 @@
 <div class="main-content">
     <div class="d-flex justify-content-between align-items-center mb-4">
         <h1><i class="fas fa-tachometer-alt me-2"></i>Dashboard</h1>
-        <a href="{{ url_for('add_transaction') }}" class="btn btn-primary">
+        <a href="{{ url_for('transact.add_transaction') }}" class="btn btn-primary">
             <i class="fas fa-plus me-1"></i>Add Transaction
         </a>
     </div>
@@ -35,12 +35,12 @@
                             </div>
                         {% endfor %}
                         <div class="mt-3">
-                            <a href="{{ url_for('accounts') }}" class="btn btn-outline-primary btn-sm">
+                            <a href="{{ url_for('acct.accounts') }}" class="btn btn-outline-primary btn-sm">
                                 <i class="fas fa-cog me-1"></i>Manage Accounts
                             </a>
                         </div>
                     {% else %}
-                        <p class="text-muted">No accounts found. <a href="{{ url_for('add_account') }}">Add your first account</a>.</p>
+                        <p class="text-muted">No accounts found. <a href="{{ url_for('acct.add_account') }}">Add your first account</a>.</p>
                     {% endif %}
                 </div>
             </div>
@@ -73,12 +73,12 @@
                             </div>
                         {% endfor %}
                         <div class="mt-3">
-                            <a href="{{ url_for('transactions') }}" class="btn btn-outline-primary btn-sm">
+                            <a href="{{ url_for('transact.transactions') }}" class="btn btn-outline-primary btn-sm">
                                 <i class="fas fa-list me-1"></i>View All Transactions
                             </a>
                         </div>
                     {% else %}
-                        <p class="text-muted">No transactions found. <a href="{{ url_for('add_transaction') }}">Add your first transaction</a>.</p>
+                        <p class="text-muted">No transactions found. <a href="{{ url_for('transact.add_transaction') }}">Add your first transaction</a>.</p>
                     {% endif %}
                 </div>
             </div>
