+++ templates/transactions.html
@@ -7,7 +7,7 @@
 <div class="main-content">
     <div class="d-flex justify-content-between align-items-center mb-4">
         <h1><i class="fas fa-exchange-alt me-2"></i>Transactions</h1>
-        <a href="{{ url_for('add_transaction') }}" class="btn btn-primary">
+        <a href="{{ url_for('transact.add_transaction') }}" class="btn btn-primary">
             <i class="fas fa-plus me-1"></i>Add Transaction
         </a>
     </div>
@@ -29,7 +29,7 @@
                             {% for transaction in transactions %}
                                 <tr>
                                     <td>
-                                        {% set edit_url = url_for('edit_transaction') %}
+                                        {% set edit_url = url_for('transact.edit_transaction') %}
                                         {% set edit_value = transaction.transactionid %}
                                         {% include 'edit_button.html' %}
                                     </td>
@@ -53,7 +53,7 @@
                     <ul class="pagination justify-content-center">
                         {% if has_prev %}
                             <li class="page-item">
-                                <a class="page-link" href="{{ url_for('transactions', p=p-1) }}">Previous</a>
+                                <a class="page-link" href="{{ url_for('transact.transactions', p=p-1) }}">Previous</a>
                             </li>
                         {% else %}
                             <li class="page-item disabled">
@@ -67,7 +67,7 @@
                         
                         {% if has_next %}
                             <li class="page-item">
-                                <a class="page-link" href="{{ url_for('transactions', p=p+1) }}">Next</a>
+                                <a class="page-link" href="{{ url_for('transact.transactions', p=p+1) }}">Next</a>
                             </li>
                         {% else %}
                             <li class="page-item disabled">
@@ -81,7 +81,7 @@
                     <i class="fas fa-exchange-alt fa-3x text-muted mb-3"></i>
                     <h5>No transactions found</h5>
                     <p class="text-muted">Start by adding your first transaction.</p>
-                    <a href="{{ url_for('add_transaction') }}" class="btn btn-primary">
+                    <a href="{{ url_for('transact.add_transaction') }}" class="btn btn-primary">
                         <i class="fas fa-plus me-1"></i>Add Transaction
                     </a>
                 </div>
